Adjustments to file names and testing files.
